Add pycodestyle to tests.
I'm choosing quite a few to ignore. Mostly because they're things scattered in the
current code, and we want it to pass. Perhaps in the future those can be updated
and the error codes removed from ignore. Most notibly are E501 and E265

E501 is line length. I don't think splitting lines is inherently more readable.
E265 is block comments. It's catching the block comment at the top of the code,
and unless someone wants to go in and add a space to every single one of these,
it's probably not a neccessary convention to check for